src/main/java/pl/edu/amu/wmi/dino/ultrapotegowanie/UltraPotegowanie.java
/*
 * To change this license header, choose License Headers in Project Properties.
 * To change this template file, choose Tools | Templates
 * and open the template in the editor.
 */
package pl.edu.amu.wmi.dino.ultrapotegowanie;

import java.util.Random;

/**
 *
 * @author bikol
 */
public class UltraPotegowanie {
    
    /**
     * Zwaraca wartość a^b.
     * @param a
     * @param b
     * @return 
     */
    public static String poteguj(String a, String b){
        int aa = Integer.parseInt(a);
        int bb = Integer.parseInt(b);
        
        return Long.toString(Math.round(Math.pow((double)aa, (double)bb)));
    }
    
}
